Extract PaymentState and ProductState types
