+++ ios/MullvadVPN/OutOfTimeViewController.swift
@@ -363,82 +363,3 @@ extension OutOfTimeViewController: RootContainment {
         false
     }
 }
-
-// MARK: - UI Restrictions
-
-private extension OutOfTimeViewController {
-    enum PaymentState: Equatable {
-        case none
-        case makingPayment(SKPayment)
-        case restoringPurchases
-
-        var allowsViewInteraction: Bool {
-            switch self {
-            case .none:
-                return true
-            case .restoringPurchases, .makingPayment:
-                return false
-            }
-        }
-    }
-
-    enum ProductState {
-        case none
-        case fetching(StoreSubscription)
-        case received(SKProduct)
-        case failed
-        case cannotMakePurchases
-
-        var isFetching: Bool {
-            if case .fetching = self {
-                return true
-            }
-            return false
-        }
-
-        var isReceived: Bool {
-            if case .received = self {
-                return true
-            }
-            return false
-        }
-
-        var purchaseButtonTitle: String? {
-            switch self {
-            case .none:
-                return nil
-
-            case let .fetching(subscription):
-                return subscription.localizedTitle
-
-            case let .received(product):
-                let localizedTitle = product.customLocalizedTitle ?? ""
-                let localizedPrice = product.localizedPrice ?? ""
-
-                let format = NSLocalizedString(
-                    "PURCHASE_BUTTON_TITLE_FORMAT",
-                    tableName: "OutOfTime",
-                    value: "%1$@ (%2$@)",
-                    comment: ""
-                )
-                return String(format: format, localizedTitle, localizedPrice)
-
-            case .failed:
-                return NSLocalizedString(
-                    "PURCHASE_BUTTON_CANNOT_CONNECT_TO_APPSTORE_LABEL",
-                    tableName: "OutOfTime",
-                    value: "Cannot connect to AppStore",
-                    comment: ""
-                )
-
-            case .cannotMakePurchases:
-                return NSLocalizedString(
-                    "PURCHASE_BUTTON_PAYMENTS_RESTRICTED_LABEL",
-                    tableName: "OutOfTime",
-                    value: "Payments restricted",
-                    comment: ""
-                )
-            }
-        }
-    }
-}
